+++ point.h
@@ -8,7 +8,7 @@ point creerPoint(signed int, signed int);
 void detruirePoint(point);
 signed int getX(point);
 signed int getY(point);
-int distanceEntreDeuxPoints(point,point);
+float distanceEntreDeuxPoints(point,point);
